refactor: move to proper static file architecture
- Remove massive HTML template string anti-pattern
- Create separate static files: index.html, style.css, app.js
- Python script now just serves static files and API endpoints
- Much cleaner separation of concerns
- No more quote escaping issues
- Easier to maintain and debug
- Static files in scripts/log_viewer_static/
