+++ scripts/log_viewer_static/style.css
@@ -1,0 +1,249 @@
+body {
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+    margin: 0;
+    padding: 0;
+    background: #1e1e1e;
+    color: #d4d4d4;
+}
+
+.container {
+    display: flex;
+    height: 100vh;
+}
+
+.sidebar {
+    width: 300px;
+    background: #252526;
+    overflow-y: auto;
+    border-right: 1px solid #3e3e42;
+}
+
+.content {
+    flex: 1;
+    overflow-y: auto;
+    padding: 20px;
+}
+
+.log-entry {
+    padding: 12px;
+    border-bottom: 1px solid #3e3e42;
+    cursor: pointer;
+    transition: background 0.2s;
+}
+
+.log-entry:hover {
+    background: #2d2d30;
+}
+
+.log-entry.selected {
+    background: #094771;
+}
+
+.log-type {
+    display: inline-block;
+    padding: 2px 8px;
+    border-radius: 3px;
+    font-size: 11px;
+    font-weight: 600;
+    text-transform: uppercase;
+    margin-right: 8px;
+}
+
+.type-request { background: #0e639c; }
+.type-response { background: #0f7b0f; }
+.type-tool_execution { background: #9f0f9f; }
+.type-error { background: #9f0f0f; }
+
+.timestamp {
+    font-size: 11px;
+    color: #858585;
+}
+
+.model-name {
+    font-size: 12px;
+    color: #4ec9b0;
+    margin-top: 4px;
+}
+
+pre {
+    background: #1e1e1e;
+    padding: 16px;
+    border-radius: 4px;
+    overflow-x: auto;
+    border: 1px solid #3e3e42;
+    white-space: pre-wrap;
+    word-wrap: break-word;
+}
+
+.json-key { color: #9cdcfe; }
+.json-string { color: #ce9178; }
+.json-number { color: #b5cea8; }
+.json-boolean { color: #569cd6; }
+.json-null { color: #569cd6; }
+
+h2 {
+    margin-top: 0;
+    color: #4ec9b0;
+}
+
+h3 {
+    color: #4ec9b0;
+    margin-top: 24px;
+    margin-bottom: 12px;
+}
+
+.filters {
+    padding: 12px;
+    background: #2d2d30;
+    border-bottom: 1px solid #3e3e42;
+}
+
+.filters select, .filters input {
+    background: #3c3c3c;
+    border: 1px solid #3e3e42;
+    color: #d4d4d4;
+    padding: 6px;
+    border-radius: 3px;
+    margin-right: 8px;
+}
+
+.stats {
+    padding: 12px;
+    background: #252526;
+    border-bottom: 1px solid #3e3e42;
+    font-size: 12px;
+}
+
+.stat-item {
+    display: inline-block;
+    margin-right: 16px;
+}
+
+.stat-label {
+    color: #858585;
+}
+
+.stat-value {
+    color: #4ec9b0;
+    font-weight: 600;
+}
+
+.refresh-btn, .view-toggle-btn {
+    background: #0e639c;
+    color: white;
+    border: none;
+    padding: 6px 12px;
+    border-radius: 3px;
+    cursor: pointer;
+    font-size: 12px;
+    margin-right: 8px;
+}
+
+.refresh-btn:hover, .view-toggle-btn:hover {
+    background: #1177bb;
+}
+
+.view-toggle-btn.active {
+    background: #1177bb;
+}
+
+.view-toggle-group {
+    margin-bottom: 16px;
+}
+
+.view-section {
+    display: none;
+    margin-bottom: 24px;
+}
+
+.view-section.active {
+    display: block;
+}
+
+.field {
+    margin-bottom: 16px;
+}
+
+.field-label {
+    color: #4ec9b0;
+    font-weight: 600;
+    margin-bottom: 4px;
+}
+
+.field-value {
+    background: #2d2d30;
+    padding: 12px;
+    border-radius: 4px;
+    border: 1px solid #3e3e42;
+}
+
+.field-value.text {
+    white-space: pre-wrap;
+    word-wrap: break-word;
+    font-family: 'Monaco', 'Menlo', 'Consolas', monospace;
+    font-size: 13px;
+    line-height: 1.6;
+}
+
+.rendered-content {
+    line-height: 1.6;
+}
+
+.rendered-content p {
+    margin: 0.5em 0;
+}
+
+.rendered-content code {
+    background: #1e1e1e;
+    padding: 2px 6px;
+    border-radius: 3px;
+    font-family: 'Monaco', 'Menlo', 'Consolas', monospace;
+    font-size: 13px;
+}
+
+.rendered-content pre {
+    margin: 0;
+}
+
+.rendered-content pre code {
+    display: block;
+    padding: 12px;
+}
+
+.array-item {
+    background: #2d2d30;
+    padding: 12px;
+    border-radius: 4px;
+    border: 1px solid #3e3e42;
+    margin-bottom: 8px;
+}
+
+.array-item-header {
+    color: #858585;
+    font-size: 11px;
+    font-weight: 600;
+    margin-bottom: 8px;
+}
+
+.collapsible {
+    cursor: pointer;
+    user-select: none;
+}
+
+.collapsible:before {
+    content: '▼ ';
+    display: inline-block;
+    transition: transform 0.2s;
+}
+
+.collapsible.collapsed:before {
+    transform: rotate(-90deg);
+}
+
+.collapsible-content {
+    margin-top: 8px;
+}
+
+.collapsible-content.collapsed {
+    display: none;
+}
